requirements.txt
# Core FastAPI and WebSocket dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-dotenv==1.0.0

# Google Generative AI
google-generativeai==0.8.3

# Audio processing
numpy==1.24.3
scipy==1.11.4
librosa==0.10.1
soundfile==0.12.1

# Async support for older Python versions
taskgroup==0.2.0; python_version < "3.11"

# Twilio webhook processing
twilio==8.10.0

# Audio format conversion
audioop-lts==0.2.1

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Additional audio utilities
pydub==0.25.1

# Other utilities
requests==2.31.0 